Widen disc offsets to 64-bit
The hardware recon reported a highest disc offset of 0xff72d4e0 on a
single-layer Super Mario Galaxy 2 disc - about 8 MB short of wrapping a
u32. The FST stores offset>>2, so (off << 2) done in u32 arithmetic
overflows once the stored value passes 0x40000000, and a dual-layer disc
is comfortably past that. Brawl and Xenoblade mods would have silently
redirected reads to garbage offsets.

FstFile::offset and RedirectSpec::discOffset are now u64, and the shift
widens before it shifts rather than after.

Found by measurement, not by reading: nothing here was failing yet.

+++ source/riivo/RiivoFst.hpp
@@ -25,7 +25,10 @@ namespace Riivo
 	struct FstFile
 	{
 		std::string path; // full path, lower-cased, leading '/', e.g. "/stage/1-1.arc"
-		u32 offset;       // disc byte offset (already un-shifted)
+		//! Disc byte offset, un-shifted. 64-bit on purpose: the stored value is
+		//! offset>>2, so a dual-layer partition addresses well past 4 GiB and a
+		//! u32 would silently wrap. A single-layer disc already reaches 0xff72d4e0.
+		u64 offset;
 		u32 length;       // file size in bytes
 		u32 index;        // FST entry index (for patching the entry in place)
 	};
